refactor: move configwriter into own module
