+++ bd-crash-handler/src/lib.rs
@@ -18,6 +18,7 @@
 #[path = "./monitor_test.rs"]
 mod tests;
 
+pub mod config_writer;
 pub mod global_state;
 
 use bd_log_primitives::{
@@ -35,7 +36,6 @@ use bd_proto::flatbuffers::report::bitdrift_public::fbs::issue_reporting::v_1::{
   Report,
   ReportType,
 };
-use bd_runtime::runtime::{BoolWatch, ConfigLoader, crash_reporting};
 use fbs::issue_reporting::v_1::root_as_report;
 use std::ffi::OsStr;
 use std::path::{Path, PathBuf};
@@ -49,7 +49,8 @@ fn test_global_init() {
 }
 
 const REPORTS_DIRECTORY: &str = "reports";
-const REPORT_CONFIG_NAME: &str = "config.csv";
+
+pub use config_writer::ConfigWriter;
 
 //
 // CrashLog
@@ -72,92 +73,6 @@ pub struct CrashLog {
 #[cfg_attr(test, mockall::automock)]
 pub trait CrashLogger: Send + Sync {
   fn log_crash(&self, report: &[u8]);
-}
-
-//
-// ConfigWriter
-//
-
-/// Checks whether config has changed, updating a persisted copy on disk if so
-/// This uses the following directory layout (within the SDK directory):
-/// - `reports/runtime.csv` - Cache of configuration flags for behavior
-pub struct ConfigWriter {
-  report_directory: PathBuf,
-  config_path: PathBuf,
-  crash_reporting_enabled_flag: BoolWatch<crash_reporting::Enabled>,
-  shutdown: bd_shutdown::ComponentShutdown,
-}
-
-impl ConfigWriter {
-  pub fn new(
-    runtime: &ConfigLoader,
-    sdk_directory: &Path,
-    shutdown: bd_shutdown::ComponentShutdown,
-  ) -> Self {
-    let report_directory = sdk_directory.join(REPORTS_DIRECTORY);
-    Self {
-      crash_reporting_enabled_flag: runtime.register_bool_watch(),
-      config_path: report_directory.join(REPORT_CONFIG_NAME),
-      report_directory,
-      shutdown,
-    }
-  }
-
-  pub async fn run(&mut self) -> anyhow::Result<()> {
-    if let Ok(exists) = tokio::fs::try_exists(self.config_path.clone()).await
-      && !exists
-    {
-      self.write_config_file().await;
-    }
-
-    self.check_for_config_changes().await
-  }
-
-  async fn check_for_config_changes(&mut self) -> anyhow::Result<()> {
-    loop {
-      tokio::select! {
-        _ = self.crash_reporting_enabled_flag.changed() => {},
-        () = self.shutdown.cancelled() => return Ok(()),
-      };
-
-      // There is chance here of the value changing during platform read, in
-      // which case crash reporting would be disabled for that session
-
-      let _ = self.crash_reporting_enabled_flag.read_mark_update();
-      self.write_config_file().await;
-    }
-  }
-
-  async fn write_config_file(&self) {
-    let crash_reporting_enabled = *self.crash_reporting_enabled_flag.read();
-    log::debug!(
-      "Writing enabled:{crash_reporting_enabled} to report config file {}",
-      self.config_path.display()
-    );
-
-    self.try_ensure_directories_exist().await;
-
-    let contents = format!("crash_reporting.enabled,{crash_reporting_enabled}");
-    if let Err(e) = tokio::fs::write(&self.config_path, contents).await {
-      log::warn!(
-        "Failed to write report directories config file: {} ({})",
-        self.config_path.display(),
-        e
-      );
-    }
-  }
-
-  async fn try_ensure_directories_exist(&self) {
-    // This can fail and we can't do anything about it, so swallow the error. Everything else needs
-    // to be resilient to the directory not existing.
-    if let Err(e) = tokio::fs::create_dir_all(self.report_directory.clone()).await {
-      log::warn!(
-        "Failed to create crash directory: {} ({})",
-        self.report_directory.display(),
-        e
-      );
-    }
-  }
 }
 
 //
